+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 v0.6.0 (in development)
 -----------------------
 - Always open text files in UTF-8
+- If Git is not installed, the Git integration will now be automatically
+  disabled instead of raising an error
+- Require Python 3.8+
+- Now available as a package on PyPI
 
 v0.5.0 (2022-07-03)
 -------------------